chore(cmd,pkg): properly preload driver version for all drivers commands.
Moreover, enforce that driver version is a semver and repos is a list of requestURI.
